requirements.txt
attrs==24.3.0
certifi==2024.12.14
h11==0.14.0
helium==5.1.0
idna==3.10
numpy==2.2.1
outcome==1.3.0.post0
pandas==2.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
